implement simple APC cache binning
Integration tests found a bug fix in isCloseEnough as well

+++ transitclock/src/main/java/org/transitclock/avl/ApcDataProcessor.java
@@ -269,7 +269,7 @@ public class ApcDataProcessor {
   }
 
 
-  public Integer getBoardingsPerMinute(String stopId, Date arrivalTime) {
+  public Double getBoardingsPerMinute(String stopId, Date arrivalTime) {
     return apcAggregator.getBoardingsPerMinute(stopId, arrivalTime);
   }
